third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/font-weight-expected.txt

This is a testharness.js-based test.
[FAIL] Can set 'font-weight' to a number: 0
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'font-weight' to a number: -3.14
  assert_approx_equals: expected -3.14 +/- 0.000001 but got -3
[FAIL] Can set 'font-weight' to a number: 3.14
  assert_approx_equals: expected 3.14 +/- 0.000001 but got 3
Harness: the test ran to completion.
